backend/requirements.txt
fastapi
uvicorn
ollama
google-genai
pydantic
requests
pyinstaller
